Add authentication
This is currently supported in buf-lint-action, buf-breaking-action and
buf-push action, but not yet in setup. Allow it here so users can setup
authentication in the setup step to use buf directly in scripts later, and so
there's a clear indication you can set it up once in buf setup and then don't
have to continue passing it for the others.