Allow configuring the buf remote
Add an input to the setup action that can control which Buf Schema Registry to login to, which will be useful for users who have a private BSR instance. I used `domain` as this is what we refer to in the cli help:
```
$ buf registry login --help
Log in to the Buf Schema Registry

This prompts for your BSR username and a BSR token and updates your .netrc file with these credentials.
The <domain> argument will default to buf.build if not specified.

Usage:
  buf registry login <domain> [flags]
```